+++ todo-app-encora/src/main/java/com/assigment/todoapp/service/ToDoService.java
@@ -2,7 +2,11 @@ package com.assigment.todoapp.service;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -57,6 +61,104 @@ public class ToDoService {
 
 	    return stream.collect(Collectors.toList());
 	}
+	
+
+
+	public Map<String, Object> paginateToDoItems(
+	            List<ToDoItem> todoItems,
+	            int page,
+	            int pageSize) {
+
+	        int totalItems = todoItems.size();
+	        int totalPages = (int) Math.ceil((double) totalItems / pageSize);
+	        int startIndex = (page - 1) * pageSize;
+	        int endIndex = Math.min(startIndex + pageSize, totalItems);
+
+	        List<ToDoItem> paginatedItems = todoItems.subList(startIndex, endIndex);
+	        int itemsOnPage = paginatedItems.size();
+
+	        Map<String, Object> response = new HashMap<>();
+	        response.put("items", paginatedItems);
+	        response.put("currentPage", page);
+	        response.put("totalItems", totalItems);
+	        response.put("totalPages", totalPages);
+	        response.put("itemsOnPage", itemsOnPage);
+
+	        return response;
+	}
+	
+	public List<ToDoItem> sortToDoItems(
+            List<ToDoItem> todoItems,
+            String sortBy1,
+            String order1,
+            String sortBy2,
+            String order2) {
+
+        Collections.sort(todoItems, new Comparator<ToDoItem>() {
+            @Override
+            public int compare(ToDoItem t1, ToDoItem t2) {
+                Map<String, Integer> priorityValues = new HashMap<>();
+                priorityValues.put("Low", 1);
+                priorityValues.put("Medium", 2);
+                priorityValues.put("High", 3);
+
+                int comparison1 = 0;
+                if (sortBy1.equals("priority")) {
+                    int priority1Value = priorityValues.getOrDefault(t1.getPriority(), 0);
+                    int priority2Value = priorityValues.getOrDefault(t2.getPriority(), 0);
+                    if (order1.equals("desc")) {
+                        priority1Value = 4 - priority1Value;
+                        priority2Value = 4 - priority2Value;
+                    }
+                    comparison1 = Integer.compare(priority1Value, priority2Value);
+                } else if (sortBy1.equals("dueDate")) {
+                    if (t1.getDueDate() == null && t2.getDueDate() == null) {
+                        comparison1 = 0;
+                    } else if (t1.getDueDate() == null) {
+                        comparison1 = -1;
+                    } else if (t2.getDueDate() == null) {
+                        comparison1 = 1;
+                    } else {
+                        comparison1 = t1.getDueDate().compareTo(t2.getDueDate());
+                    }
+                    if (order1.equals("desc")) {
+                        comparison1 = -comparison1;
+                    }
+                }
+
+                if (comparison1 != 0) {
+                    return comparison1;
+                } else {
+                    int comparison2 = 0;
+                    if (sortBy2.equals("priority")) {
+                        int priority1Value = priorityValues.getOrDefault(t1.getPriority(), 0);
+                        int priority2Value = priorityValues.getOrDefault(t2.getPriority(), 0);
+                        if (order2.equals("desc")) {
+                            priority1Value = 4 - priority1Value;
+                            priority2Value = 4 - priority2Value;
+                        }
+                        comparison2 = Integer.compare(priority1Value, priority2Value);
+                    } else if (sortBy2.equals("dueDate")) {
+                        if (t1.getDueDate() == null && t2.getDueDate() == null) {
+                            comparison2 = 0;
+                        } else if (t1.getDueDate() == null) {
+                            comparison2 = -1;
+                        } else if (t2.getDueDate() == null) {
+                            comparison2 = 1;
+                        } else {
+                            comparison2 = t1.getDueDate().compareTo(t2.getDueDate());
+                        }
+                        if (order2.equals("desc")) {
+                            comparison2 = -comparison2;
+                        }
+                    }
+                    return comparison2;
+                }
+            }
+        });
+
+        return todoItems;
+    }
 
 	public ToDoItem createToDoItem(ToDoItem todoItem) {
         todoItem.setId(UUID.randomUUID());
